fix(stats): Show stats when selecting tubes
Previously, there was a panic when attempting to show stats for tubes
that did not include the default tube. Now, do a check to see if the
default tube is among those selected.
Also update the function to sort map keys to use new functions.
Update signature of PrintStats for testing purposes.

Fixes the issue described at src-d/beanstool#17